requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.13.9
botocore==1.16.9
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
pandas==1.0.3
pathlib==1.0.1
Pillow==7.1.2
psycopg2==2.8.5
public==2019.4.13
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
temp==2020.4.16
toml==0.10.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
